+++ prefect_lakefs/refs.py
@@ -1,13 +1,7 @@
 """Module for interacting with LakeFS refs api from Prefect flows."""
 from typing import Any, Dict
 
-from lakefs_client.models import (
-    CommitList,
-    DiffList,
-    FindMergeBaseResult,
-    MergeResult,
-    RefsDump,
-)
+from lakefs_client.models import CommitList, DiffList, FindMergeBaseResult, MergeResult
 from prefect import task
 from prefect.utilities.asyncutils import run_sync_in_worker_thread
 
@@ -60,48 +54,6 @@ async def diff_refs(
             repository=repository,
             left_ref=left_ref,
             right_ref=right_ref,
-            **lakefs_kwargs,
-        )
-
-
-@task
-async def dump_refs(
-    repository: str,
-    lakefs_credentials: LakeFSCredentials,
-    **lakefs_kwargs: Dict[str, Any],
-) -> RefsDump:
-    """Dump repository refs (tags, commits, branches) to object store.
-
-    Args:
-        lakefs_credentials: `LakeFSCredentials` block for creating
-            authenticated LakeFS API clients.
-        repository: name of a lakefs repository.
-        **lakefs_kwargs: Optional extra keyword arguments to pass to the LakeFS API.
-
-    Returns:
-        A LakeFS `RefsDump` object.
-
-    Example:
-        dump refs for repository named `example`:
-
-        ```python
-        from prefect import flow
-        from prefect_lakefs import LakeFSCredentials
-        from prefect_lakefs.tasks import dump_refs
-
-        @flow
-        def dump_refs_example_repo():
-            dump = dump_refs(
-                lakefs_credentials=LakeFSCredentials.load("lakefs-creds"),
-                repository="example",
-            )
-        ```
-    """
-
-    with lakefs_credentials.get_client("refs") as refs:
-        return await run_sync_in_worker_thread(
-            refs.dump_refs,
-            repository=repository,
             **lakefs_kwargs,
         )
 
@@ -248,56 +200,5 @@ async def merge_into_branch(
             repository=repository,
             source_ref=source_ref,
             destination_branch=destination_branch,
-            **lakefs_kwargs,
-        )
-
-
-@task
-async def restore_refs(
-    repository: str,
-    refs_dump: RefsDump,
-    lakefs_credentials: LakeFSCredentials,
-    **lakefs_kwargs: Dict[str, Any],
-) -> None:
-    """Restore repository refs form object store.
-
-    Args:
-        lakefs_credentials: `LakeFSCredentials` block for creating
-            authenticated LakeFS API clients.
-        repository: name of a lakefs repository.
-        refs_dump: LakeFS RefsDump Model.
-        **lakefs_kwargs: Optional extra keyword arguments to pass to the LakeFS API.
-
-    Returns:
-        None
-
-    Example:
-        restore refs dump for repository named `example`:
-
-        ```python
-        from prefect import flow
-        from prefect_lakefs import LakeFSCredentials
-        from prefect_lakefs.tasks import dump_refs
-        from prefect_lakefs.tasks import restore_refs
-
-        @flow
-        def restore_refs_example_repo():
-            dump = dump_refs(
-                lakefs_credentials=LakeFSCredentials.load("lakefs-creds"),
-                repository="example",
-            )
-            restore_refs(
-                lakefs_credentials=LakeFSCredentials.load("lakefs-creds"),
-                repository="example",
-                refs_dump=dump,
-            )
-        ```
-    """
-
-    with lakefs_credentials.get_client("refs") as refs:
-        return await run_sync_in_worker_thread(
-            refs.restore_refs,
-            repository=repository,
-            refs_dump=refs_dump,
             **lakefs_kwargs,
         )
