+++ requirements.txt
@@ -1,4 +1,5 @@
 prefect>=2.0.0
 prefect-aws>=0.2.4
-lakefs-client>=0.96
+lakefs-client>=0.113.0
 gitpython>=3.1.31
+pydantic<2
