+++ tests/test_refs.py
@@ -1,12 +1,8 @@
-from lakefs_client.models import RefsDump
-
 from prefect_lakefs.refs import (
     diff_refs,
-    dump_refs,
     find_merge_base,
     log_commits,
     merge_into_branch,
-    restore_refs,
 )
 
 
@@ -20,14 +16,6 @@ async def test_diff_refs(lakefs_credentials, _mock_refs_client):
     assert _mock_refs_client.diff_refs.call_args[1]["repository"] == "example"
     assert _mock_refs_client.diff_refs.call_args[1]["left_ref"] == "left"
     assert _mock_refs_client.diff_refs.call_args[1]["right_ref"] == "right"
-
-
-async def test_dump_refs(lakefs_credentials, _mock_refs_client):
-    await dump_refs.fn(
-        repository="example",
-        lakefs_credentials=lakefs_credentials,
-    )
-    assert _mock_refs_client.dump_refs.call_args[1]["repository"] == "example"
 
 
 async def test_find_merge_base(lakefs_credentials, _mock_refs_client):
@@ -66,16 +54,3 @@ async def test_merge_into_branch(lakefs_credentials, _mock_refs_client):
     assert (
         _mock_refs_client.merge_into_branch.call_args[1]["destination_branch"] == "dest"
     )
-
-
-async def test_restore_refs(lakefs_credentials, _mock_refs_client):
-    await restore_refs.fn(
-        repository="example",
-        refs_dump=RefsDump(
-            commits_meta_range_id="commits_meta_range_id",
-            tags_meta_range_id="tags_meta_range_id",
-            branches_meta_range_id="branches_meta_range_id",
-        ),
-        lakefs_credentials=lakefs_credentials,
-    )
-    assert _mock_refs_client.restore_refs.call_args[1]["repository"] == "example"
